Remove unneeded code
