+++ routes/api/timetable.js
@@ -1,6 +1,7 @@
 const express = require("express");
 const router = express.Router();
-const { check, validationResult } = require("express-validator");
+const { check, query, validationResult } = require("express-validator");
+
 
 const loggedIn = require("../../middleware/auth");
 
@@ -81,18 +82,40 @@ router.get("/toggleShare/:id", loggedIn, async (req, res) => {
   }
 });
 
-router.get("/viewShared/:id", loggedIn, async (req, res) => {
-  try {
-    const ownerId = await Student.findOne({
-      email: req.params.id.append("@hyderabad.bits-pilani.ac.in")
-    });
-    const tt = await TimeTable.find({
-      ownerId: ownerId
-    });
-    res.status(200).json(tt);
-  } catch (err) {
-    res.status(500).send("Server error");
-  }
-});
+router.get("/viewshared",
+    [
+        loggedIn,
+        query("branch").isArray(),
+        query("year").not()
+            .isEmpty()
+    ],
+    async (req, res) => {
+        const errors = validationResult(req);
+        if (!errors.isEmpty()) {
+            res.status(422).json({ errors: errors.array() });
+        }
+        else {
+
+            try {
+                await TimeTable.find({ isShared: true }).populate({ path: 'ownerId', match: { 'branch': req.query.branch, 'year': parseInt(req.query.year) }, select: 'name' }).exec(
+                    (error, docs) => {
+                        if (error) { res.status(500).send(err); }
+                        else {
+                            let TTList = [];
+                            for (let TT of docs) {
+                                TTList.push(TT);
+                            }
+                            res.json(TTList);
+                        }
+                    })
+            }
+            catch (err) {
+
+                console.error(err.message);
+                res.status(500).send("Server error");
+            }
+        }
+    })
+
 
 module.exports = router;
